requirements.txt
openai
google-generativeai
groq
mistralai

langchain
langchain-core
langchain-openai
langchain-google-genai
langchain-community
langchain-groq
langchain-mistralai

python-dotenv
fastapi
uvicorn
faiss-cpu
tiktoken

pydantic
requests
tqdm